Watson working

+++ db/migrate/20160826140510_create_hosts.rb
@@ -5,12 +5,12 @@ class CreateHosts < ActiveRecord::Migration[5.0]
       t.string :bio
       t.string :address
       t.string :city_state
-      t.integer :host_expectations
-      t.integer :openness
-      t.integer :conscientiousness
-      t.integer :extraversion
-      t.integer :agreeableness
-      t.integer :emotional_range
+      t.float :host_expectations
+      t.float :openness
+      t.float :conscientiousness
+      t.float :extraversion
+      t.float :agreeableness
+      t.float :emotional_range
       t.string :profile_link
 
       t.timestamps
